Functions which return vector.size() now correctly return size_t

+++ gemcutter/Resource/Texture.cpp
@@ -448,7 +448,7 @@ namespace gem
 		textureSlots.clear();
 	}
 
-	unsigned TextureList::Size() const
+	size_t TextureList::Size() const
 	{
 		return textureSlots.size();
 	}
